@@ -3,14 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^3.9.2",
-    "rc-tooltip": "^3.7.3",
-    "react": "^16.8.3",
-    "react-calendar-heatmap": "^1.8.0",
-    "react-dom": "^16.8.3",
-    "react-scripts": "2.1.5",
-    "react-social-icons": "^4.1.0",
-    "react-tooltip": "^3.9.2"
+    "@emotion/react": "^11.10.8",
+    "@emotion/styled": "^11.10.8",
+    "@mui/material": "^5.12.3",
+    "rc-tooltip": "^6.0.1",
+    "react": "^18.2.0",
+    "react-calendar-heatmap": "^1.9.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "react-social-icons": "^5.15.0",
+    "react-tooltip": "^5.11.2"
   },
   "scripts": {
     "start": "react-scripts start",
